# backend/SwaggerDocumentation.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Portal Web UCB API",
    "version": "1.0.0",
    "description": "La API de la Universidad Católica Boliviana es un conjunto de endpoints que proporcionan acceso a diversas funcionalidades de tu portal web. Esta API permite a los usuarios interactuar con recursos como noticias, usuarios, autenticación, y otras características clave de la plataforma. Some useful links:  - [Repositorio](https://github.com/mcvagut34/proytecnoweb)"
  },
  "servers": [
    {
      "url": "http://localhost:8800"
    }
  ],
  "paths": {
    "/api/autenticacion/login": {
      "post": {
        "summary": "Iniciar sesión como administrador",
        "description": "Inicia sesión como administrador utilizando credenciales.",
        "tags": [
          "Autenticación"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inicio de sesión exitoso"
          },
          "401": {
            "description": "Credenciales incorrectas"
          }
        }
      }
    },
    "/api/autenticacion/logout": {
      "post": {
        "summary": "Cerrar sesión como administrador",
        "description": "Cierra la sesión de administrador.",
        "tags": [
          "Autenticación"
        ],
        "responses": {
          "200": {
            "description": "Sesión cerrada con éxito"
          }
        }
      }
    },
    "/api/autenticacion/registro": {
      "post": {
        "summary": "Registrar usuario",
        "description": "Registra un nuevo usuario como administrador.",
        "tags": [
          "Autenticación"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario registrado con éxito"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      }
    },
    "/api/noticias/crear": {
      "post": {
        "summary": "Crear una noticia",
        "description": "Crea una nueva noticia.",
        "tags": [
          "Noticias"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Noticia"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Noticia creada con éxito"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      }
    },
    "/api/noticias": {
      "get": {
        "summary": "Obtener noticias",
        "description": "Obtiene una lista de noticias.",
        "tags": [
          "Noticias"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de noticias obtenida con éxito"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      }
    },
    "/api/noticias/{id}": {
      "get": {
        "summary": "Obtener noticia por ID",
        "description": "Obtiene una noticia por su ID.",
        "tags": [
          "Noticias"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID de la noticia a obtener",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Noticia obtenida con éxito"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      },
      "put": {
        "summary": "Actualizar noticia por ID",
        "description": "Actualiza una noticia por su ID.",
        "tags": [
          "Noticias"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID de la noticia a actualizar",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Noticia actualizada con éxito"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      },
      "patch": {
        "summary": "Borrar noticia por ID",
        "description": "Borra una noticia por su ID.",
        "tags": [
          "Noticias"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID de la noticia a borrar",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Noticia borrada con éxito"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      }
    },
    "/api/usuarios/{id}": {
      "put": {
        "summary": "Actualizar usuario por ID",
        "description": "Actualiza un usuario por su ID.",
        "tags": [
          "Usuarios"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID del usuario a actualizar",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario actualizado con éxito"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      },
      "patch": {
        "summary": "Borrar usuario por ID",
        "description": "Borra un usuario por su ID.",
        "tags": [
          "Usuarios"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID del usuario a borrar",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario borrado con éxito"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      },
      "get": {
        "summary": "Obtener usuario por ID",
        "description": "Obtiene un usuario por su ID.",
        "tags": [
          "Usuarios"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID del usuario a obtener",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario obtenido con éxito"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      }
    },
    "/api/usuarios": {
      "get": {
        "summary": "Obtener todos los usuarios",
        "description": "Obtiene una lista de todos los usuarios.",
        "tags": [
          "Usuarios"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios obtenida con éxito"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
    "schemas": {
      "Usuario": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "description": "Nombre del usuario"
          },
          "apellido": {
            "type": "string",
            "description": "Apellido del usuario"
          },
          "fechaNacimiento": {
            "type": "string",
            "format": "date",
            "description": "Fecha de nacimiento del usuario"
          },
          "usuario": {
            "type": "string",
            "description": "Nombre de usuario"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Correo electrónico del usuario"
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "Contraseña del usuario"
          },
          "pais": {
            "type": "string",
            "description": "País del usuario"
          },
          "img": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Imágenes del usuario"
          },
          "ciudad": {
            "type": "string",
            "description": "Ciudad del usuario"
          },
          "telefono": {
            "type": "string",
            "description": "Número de teléfono del usuario"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Roles del usuario"
          },
          "activo": {
            "type": "boolean",
            "description": "Estado de activación del usuario"
          },
          "intentosFallidos": {
            "type": "integer",
            "description": "Número de intentos de inicio de sesión fallidos"
          },
          "resetPasswordToken": {
            "type": "string",
            "description": "Token para restablecer la contraseña"
          },
          "resetPasswordExpires": {
            "type": "string",
            "format": "date",
            "description": "Fecha de vencimiento del token de restablecimiento de contraseña"
          }
        },
        "required": [
          "nombre",
          "apellido",
          "fechaNacimiento",
          "usuario",
          "email",
          "password",
          "pais",
          "ciudad",
          "roles",
          "activo",
          "intentosFallidos",
          "resetPasswordToken",
          "resetPasswordExpires"
        ]
      },
      "Noticia": {
        "type": "object",
        "properties": {
          "titulo": {
            "type": "string",
            "description": "Título de la noticia"
          },
          "descripcion": {
            "type": "string",
            "description": "Descripción de la noticia"
          },
          "destacada": {
            "type": "boolean",
            "description": "Indica si la noticia está destacada"
          },
          "activo": {
            "type": "boolean",
            "description": "Estado de activación de la noticia"
          },
          "contenido": {
            "type": "string",
            "description": "Contenido de la noticia"
          },
          "fecha_publicacion": {
            "type": "string",
            "format": "date",
            "description": "Fecha de publicación de la noticia"
          },
          "img": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Imágenes asociadas a la noticia"
          }
        },
        "required": [
          "titulo",
          "descripcion",
          "contenido"
        ]
      }
    }
  },
  "tags": []
}